refactor(svelte2tsx): template のウォーク状態を ctx.rs へ切り出す
TemplateNodeExt / Counter / 要素オープナーのコメント範囲 thread_local を
template/ctx.rs へ純粋移動。出力・ロジックの変更なし。
