+++ crates/rsvelte_core/src/svelte2tsx/template/mod.rs
@@ -6,6 +6,8 @@
 //! Each template node type has a corresponding handler that overwrites the
 //! original source range with the appropriate TypeScript/TSX code.
 
+mod ctx;
+
 use crate::ast::template::{
     AttachTag, Attribute, AttributeNode, AttributeValue, AttributeValuePart, AwaitBlock,
     BindDirective, ClassDirective, Comment, Component, ConstTag, DebugTag, EachBlock,
@@ -20,6 +22,9 @@ use indexmap::IndexMap;
 
 use super::magic_string::MagicString;
 use super::svelte2tsx::{Svelte2TsxOptions, SvelteVersion, slice_src};
+use ctx::{Counter, ELEMENT_OPENER_COMMENTS, TemplateNodeExt};
+
+pub(crate) use ctx::{clear_element_opener_comments, set_element_opener_comments};
 
 // =============================================================================
 // Template context for collecting slot/event information
@@ -51,84 +56,6 @@ pub enum ForwardedEventKind {
     /// Component / `svelte:component` — official `handleEventHandlerBubble`
     /// concats into the existing entry.
     Component,
-}
-
-// =============================================================================
-// TemplateNode position helpers
-// =============================================================================
-
-/// Extension trait for getting start/end positions from TemplateNode.
-trait TemplateNodeExt {
-    fn start(&self) -> u32;
-    fn end(&self) -> u32;
-}
-
-impl TemplateNodeExt for TemplateNode<'_> {
-    fn start(&self) -> u32 {
-        match self {
-            TemplateNode::Text(n) => n.start,
-            TemplateNode::Comment(n) => n.start,
-            TemplateNode::TitleElement(n) => n.start,
-            TemplateNode::SlotElement(n) => n.start,
-            TemplateNode::SvelteBody(n)
-            | TemplateNode::SvelteDocument(n)
-            | TemplateNode::SvelteFragment(n)
-            | TemplateNode::SvelteBoundary(n)
-            | TemplateNode::SvelteHead(n)
-            | TemplateNode::SvelteOptions(n)
-            | TemplateNode::SvelteSelf(n)
-            | TemplateNode::SvelteWindow(n) => n.start,
-            TemplateNode::ExpressionTag(n) => n.start,
-            TemplateNode::HtmlTag(n) => n.start,
-            TemplateNode::ConstTag(n) => n.start,
-            TemplateNode::DeclarationTag(n) => n.start,
-            TemplateNode::DebugTag(n) => n.start,
-            TemplateNode::RenderTag(n) => n.start,
-            TemplateNode::AttachTag(n) => n.start,
-            TemplateNode::IfBlock(n) => n.start,
-            TemplateNode::EachBlock(n) => n.start,
-            TemplateNode::AwaitBlock(n) => n.start,
-            TemplateNode::KeyBlock(n) => n.start,
-            TemplateNode::SnippetBlock(n) => n.start,
-            TemplateNode::RegularElement(n) => n.start,
-            TemplateNode::Component(n) => n.start,
-            TemplateNode::SvelteComponent(n) => n.start,
-            TemplateNode::SvelteElement(n) => n.start,
-        }
-    }
-
-    fn end(&self) -> u32 {
-        match self {
-            TemplateNode::Text(n) => n.end,
-            TemplateNode::Comment(n) => n.end,
-            TemplateNode::TitleElement(n) => n.end,
-            TemplateNode::SlotElement(n) => n.end,
-            TemplateNode::SvelteBody(n)
-            | TemplateNode::SvelteDocument(n)
-            | TemplateNode::SvelteFragment(n)
-            | TemplateNode::SvelteBoundary(n)
-            | TemplateNode::SvelteHead(n)
-            | TemplateNode::SvelteOptions(n)
-            | TemplateNode::SvelteSelf(n)
-            | TemplateNode::SvelteWindow(n) => n.end,
-            TemplateNode::ExpressionTag(n) => n.end,
-            TemplateNode::HtmlTag(n) => n.end,
-            TemplateNode::ConstTag(n) => n.end,
-            TemplateNode::DeclarationTag(n) => n.end,
-            TemplateNode::DebugTag(n) => n.end,
-            TemplateNode::RenderTag(n) => n.end,
-            TemplateNode::AttachTag(n) => n.end,
-            TemplateNode::IfBlock(n) => n.end,
-            TemplateNode::EachBlock(n) => n.end,
-            TemplateNode::AwaitBlock(n) => n.end,
-            TemplateNode::KeyBlock(n) => n.end,
-            TemplateNode::SnippetBlock(n) => n.end,
-            TemplateNode::RegularElement(n) => n.end,
-            TemplateNode::Component(n) => n.end,
-            TemplateNode::SvelteComponent(n) => n.end,
-            TemplateNode::SvelteElement(n) => n.end,
-        }
-    }
 }
 
 // =============================================================================
@@ -460,55 +387,6 @@ fn reversed_component_instance_name(name: &str, depth: u32) -> String {
     let sanitized = sanitize_prop_name(name);
     let reversed: String = sanitized.chars().rev().collect();
     format!("$$_{}{}", reversed, depth)
-}
-
-/// Counter for generating unique variable names.
-/// Uses per-name counters so each unique component/element name gets its own counter.
-struct Counter {
-    counters: std::collections::HashMap<String, u32>,
-    /// When set (to a component instance var), a `slot="name"` element/component
-    /// encountered while processing that component's children — at any depth
-    /// inside `{#each}`/`{#if}`/etc. control-flow blocks — is lowered to the
-    /// named-slot `$$slot_def[...]` form referencing this instance var. Cleared
-    /// when descending into a nested element/component (which owns its own slot
-    /// scope). Threaded via `&mut Counter` so the 30+ existing
-    /// `process_*_inplace` call sites need no signature change.
-    slot_inst: Option<String>,
-    /// Set just before `handle_named_slot_component` calls `handle_component`:
-    /// a component that is a named-slot child has its component-name reference
-    /// (`Inner;`) emitted by the caller *outside* the component's own block
-    /// (between the component-block close and the named-slot-block close). So
-    /// `handle_component` closes its block with a bare `}` (no name) and the
-    /// caller emits ` Name}`. Mirrors official `endTransformation` ordering
-    /// `['}'(slotLet), name, '}']`. Consumed once at the top of `handle_component`.
-    named_slot_component_close: bool,
-    /// Set just before `handle_named_slot_component` calls `handle_component`:
-    /// a component that is a named-slot child (`<C slot="x" let:y>`) has its
-    /// `let:` directives consumed by the parent's `$$slot_def["x"]` destructure,
-    /// so `handle_component` must NOT re-emit them as the component's own
-    /// default-slot let block. Consumed once at the top of `handle_component`.
-    suppress_component_lets: bool,
-}
-
-impl Counter {
-    fn new() -> Self {
-        Self {
-            counters: std::collections::HashMap::new(),
-            slot_inst: None,
-            named_slot_component_close: false,
-            suppress_component_lets: false,
-        }
-    }
-    #[allow(dead_code)]
-    fn next(&mut self) -> u32 {
-        self.next_for("")
-    }
-    fn next_for(&mut self, name: &str) -> u32 {
-        let entry = self.counters.entry(name.to_string()).or_insert(0);
-        let v = *entry;
-        *entry += 1;
-        v
-    }
 }
 
 // =============================================================================
@@ -5847,24 +5725,6 @@ fn transform_attribute_case(name: &str, tag: &str, is_element: bool) -> String {
     } else {
         name.to_string()
     }
-}
-
-thread_local! {
-    /// Source ranges of comments found inside element opening tags (between
-    /// attributes), set per-compile so attribute emission can re-attach them as
-    /// leading comments. Mirrors official `attr.leadingComments`.
-    static ELEMENT_OPENER_COMMENTS: std::cell::RefCell<Vec<(u32, u32)>> =
-        const { std::cell::RefCell::new(Vec::new()) };
-}
-
-/// Set the element-opener comment ranges for the current compile (read-only).
-pub(crate) fn set_element_opener_comments(ranges: Vec<(u32, u32)>) {
-    ELEMENT_OPENER_COMMENTS.with(|c| *c.borrow_mut() = ranges);
-}
-
-/// Clear the element-opener comment ranges after a compile.
-pub(crate) fn clear_element_opener_comments() {
-    ELEMENT_OPENER_COMMENTS.with(|c| c.borrow_mut().clear());
 }
 
 /// Build the leading-comment prefix segs for an attribute starting at
